Use Biome for linting and formatting
[Biome][1] is a next-generation [opinionated][2] formatter and linter.
It is extremely fast and by using one tool for both formatting and
linting we avoid the [conflicts][3] that can often arise when using
Prettier and ESLint.

This commit installs Biome as per the [installation docs][4].
It also adds [the VS Code Biome extension][5] to our devcontainer
(Codespace), and sets it up to format files and organise imports,
automatically on save.

[1]: https://biomejs.dev/
[2]: https://biomejs.dev/formatter/option-philosophy/
[3]: https://dev.to/studio_m_song/how-to-make-eslint-work-with-prettier-avoiding-conflicts-and-problems-57pi
[4]: https://biomejs.dev/guides/getting-started/#installation
[5]: https://biomejs.dev/reference/vscode/